improved responsive part
