+++ src/App.jsx
@@ -1,13 +1,16 @@
 import './App.css';
 
-import routes from './routes';
-import { BrowserRouter} from 'react-router-dom';
-import Footer from './components/Footer';
-
+import Navbar from './components/Navbar/Navbar';
+import { BrowserRouter } from 'react-router-dom';
+import Footer from './components/Footer/Footer';
 
 function App() {
   return (
     <div className="App">
+      <BrowserRouter>
+        <Navbar isLogin />
+        <Footer />
+      </BrowserRouter>
     </div>
   );
 }
