+++ src/components/Footer/FooterResourcesLinks.js
@@ -1,0 +1,18 @@
+export default function FooterResourcesLinks() {
+  return (
+    <div className=" lg:lg:text-[15px] lg:text-[#FFFFFF] sm:invisible md:invisible lg:visible ">
+      <h5 className="font-bold mb-1">Resources</h5>
+      <ul className="list-none mb-0">
+        <li>
+          <a href="#!">Privacy Policy</a>
+        </li>
+        <li>
+          <a href="#!">Terms and Conditions</a>
+        </li>
+        <li>
+          <a href="#!">Contact Us</a>
+        </li>
+      </ul>
+    </div>
+  );
+}
