--- src/components/FooterResourcesLinks.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-export default function FooterResourcesLinks() {
-  return (
-    <div className="mt-6 text-sm text-[#FFFFFF] hidden md:block md:ml-[50px]">
-        <h5 className="font-bold mb-1">Resources</h5>
-        <ul className="list-none mb-0">
-          <li>
-            <a href="#!">Privacy Policy</a>
-          </li>
-          <li>
-            <a href="#!">Terms and Conditions</a>
-          </li>
-          <li>
-            <a href="#!">Contact Us</a>
-          </li>
-        </ul>
-      </div>
-  )
-}
