+++ src/index.js
@@ -1,22 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import 'tw-elements';
 import './index.css';
 import { Provider } from 'react-redux';
 import { store } from './app/store';
-import * as serviceWorker from './serviceWorker';
 
-import Footer from './components/Footer';
+import Footer from './components/Footer/Footer';
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-    <Footer />
+      <Footer />
     </Provider>
   </React.StrictMode>,
   document.getElementById('root'),
 );
-
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
