+++ tailwind.config.js
@@ -1,6 +1,13 @@
 module.exports = {
   content: ['./src/**/*.{js,jsx,}'],
   theme: {
+    screens: {
+      xs: '300px',
+      sm: '640px',
+      md: '768px',
+      lg: '1024px',
+      xl: '1280px',
+    },
     fontSize: {
       sm: ['14px', '20px'],
       base: ['16px', '24px'],
@@ -9,14 +16,14 @@ module.exports = {
     },
     extend: {},
     colors: {
-      'blue': '#1fb6ff',
-      'pink': '#ff49db',
-      'orange': '#FF7338',
-      'green': '#13ce66',
+      blue: '#1fb6ff',
+      pink: '#ff49db',
+      orange: '#FF7338',
+      green: '#13ce66',
       'gray-dark': '#273444',
-      'gray': '#8492a6',
+      gray: '#8492a6',
       'gray-light': '#d3dce6',
     },
   },
   plugins: [],
-}
+};
